chore: extract viewer processors into JS module
