+++ kaiserlift/viewers.py
@@ -4,6 +4,7 @@ import matplotlib.pyplot as plt
 import base64
 import re
 from io import BytesIO
+from pathlib import Path
 from .df_processers import (
     calculate_1rm,
     highest_weight_per_rep,
@@ -149,6 +150,7 @@ def print_oldest_exercise(
 def gen_html_viewer(df):
     df_records = highest_weight_per_rep(df)
     df_targets = df_next_pareto(df_records)
+    processors_js = (Path(__file__).with_name("js") / "processors.js").read_text()
 
     # Create a dictionary: { exercise_name: base64_image_string }
     figures_html: dict[str, str] = {}
@@ -210,7 +212,8 @@ def gen_html_viewer(df):
 
     # JS and CSS for DataTables + filtering
     # JS, CSS, and styling improvements
-    js_and_css = """
+    js_and_css = (
+        """
     <!-- DataTables -->
     <link rel=\"stylesheet\" href=\"https://cdn.datatables.net/1.13.4/css/jquery.dataTables.min.css\"/>
     <script src=\"https://code.jquery.com/jquery-3.5.1.js\"></script>
@@ -254,69 +257,9 @@ def gen_html_viewer(df):
     </style>
 
     <script>
-    function calculate1RM(weight, reps) {
-        weight = parseFloat(weight);
-        reps = parseInt(reps);
-        if (!weight || !reps || reps <= 0 || weight < 0) {
-            if (weight === 0 && reps > 0) { return 0; }
-            return NaN;
-        }
-        if (reps === 1) { return weight; }
-        return weight * (1 + reps / 30.0);
-    }
-
-    function highestWeightPerRep(data) {
-        const groups = {};
-        data.forEach(row => {
-            const ex = row.Exercise;
-            if (!groups[ex]) { groups[ex] = []; }
-            groups[ex].push(row);
-        });
-        const result = [];
-        Object.values(groups).forEach(rows => {
-            const byRep = {};
-            rows.forEach(r => {
-                const reps = parseInt(r.Reps);
-                const weight = parseFloat(r.Weight);
-                if (!byRep[reps] || weight > byRep[reps].Weight) {
-                    byRep[reps] = { ...r, Reps: reps, Weight: weight };
-                }
-            });
-            const candidates = Object.values(byRep);
-            candidates.forEach(c => {
-                const superseded = candidates.some(o => o.Reps > c.Reps && o.Weight >= c.Weight);
-                if (!superseded) { result.push(c); }
-            });
-        });
-        return result;
-    }
-
-    function dfNextPareto(records) {
-        const groups = {};
-        records.forEach(r => {
-            const ex = r.Exercise;
-            if (!groups[ex]) { groups[ex] = []; }
-            groups[ex].push(r);
-        });
-        const rows = [];
-        Object.entries(groups).forEach(([ex, arr]) => {
-            arr.sort((a, b) => a.Reps - b.Reps);
-            const ws = arr.map(r => r.Weight);
-            const rs = arr.map(r => r.Reps);
-            rows.push({ Exercise: ex, Weight: ws[0] + 5, Reps: 1 });
-            for (let i = 0; i < rs.length - 1; i++) {
-                if (rs[i + 1] > rs[i] + 1) {
-                    const nr = rs[i] + 1;
-                    const c1 = ws[i];
-                    const c2 = ws[i + 1] + 5;
-                    rows.push({ Exercise: ex, Weight: Math.min(c1, c2), Reps: nr });
-                }
-            }
-            rows.push({ Exercise: ex, Weight: ws[ws.length - 1], Reps: rs[rs.length - 1] + 1 });
-        });
-        rows.forEach(r => { r["1RM"] = calculate1RM(r.Weight, r.Reps); });
-        return rows;
-    }
+    """
+        + processors_js
+        + """
 
     $(document).ready(function() {
         // Initialize DataTable
@@ -377,6 +320,7 @@ def gen_html_viewer(df):
     });
     </script>
     """
+    )
 
     # Final combo
     full_html = js_and_css + dropdown_html + table_html + all_figures_html
